refactor to not use combo-box
