+++ integration/package.json
@@ -43,7 +43,6 @@
     "@vaadin/time-picker": "24.4.0-alpha22",
     "@vaadin/tooltip": "24.4.0-alpha22",
     "@vaadin/vaadin-themable-mixin": "24.4.0-alpha22",
-    "lit": "^3.0.0",
     "sinon": "^13.0.2"
   }
 }
